# docs/backend-contracts/contract-history/1.3.0-contract-914084874afda2481813a041da4cc01249aa9ea557d9a8bf29baeed4f10e0dc9/openapi.snapshot.yaml
openapi: 3.1.0
info:
  title: BNBU Sports Unified API
  version: 1.3.0-contract
  summary: Unified contract for the sports check-in platform
  description: |
    Stage 21B contract. It does not prove that every registered capability is
    business-enabled or production-ready.
    Business fields come from 02-data-dictionary.md, state transitions from
    03-state-machines.md, rules from 04-business-rules.md, and authorization
    from 05-permission-matrix.md. Foundation authentication/session persistence
    is approved; CourseInvite/Join Capability are contract-and-persistence-design
    only until the Enrollment Gate, while every Export operation is fail-closed
    and V1 persists no ExportJob. Of the 30 client-support operations added in
    Stage 21, authentication, exemption, release-policy, notification,
    push-device, preference, help, and feedback capabilities are approved for
    real local implementation, while unresolved capabilities remain explicit
    routed denials. GPS has persistence and
    application-layer foundations, but all six location HTTP operations remain
    denied. None of these statements proves Staging, iOS-binary, push-delivery,
    GPS-production, or Production-Gate readiness.
servers:
  - url: /api/v1
    description: Version 1 relative base path
security:
  - bearerAuth: []
...
paths:
...
  /system-mode:
    get:
      tags: [System Mode]
      operationId: getSystemMode
      x-access-policy:
        policyId: PUBLIC-SYSTEM-MODE-READ
        authentication: PUBLIC
        allowedRoles: []
        organizationScope: NONE
        resourceScope: NONE
        resourceResolver: NONE
        defaultDeny: true
      summary: Read the safe public SystemMode projection
      security: []
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/SystemModeSuccess'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /me:
    get:
      tags: [Current User]
      operationId: getCurrentUser
      x-access-policy:
        policyId: USER-SELF-READ
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: SELF
        resourceResolver: PRINCIPAL_USER
        defaultDeny: true
      summary: Get the authenticated User and role-specific Profile projection
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
      responses:
        '200':
          $ref: '#/components/responses/CurrentUserSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Current User]
      operationId: updateCurrentUserProfile
      x-access-policy:
        policyId: USER-SELF-UPDATE
        authentication: ACCESS_TOKEN
        allowedRoles: [STUDENT, TEACHER, ADMIN]
        organizationScope: PRINCIPAL_ORGANIZATION
        resourceScope: SELF
        resourceResolver: PRINCIPAL_USER
        defaultDeny: true
      summary: Request a verified change to the authenticated user's contact method
      description: Transport placeholder only; always returns SYSTEM_MODE_UNSUPPORTED until a verified contact-change flow is approved.
      x-enabled-by-default: false
      x-business-blocker: CONTACT_VERIFICATION_GATE
      x-default-deny-error: SYSTEM_MODE_UNSUPPORTED
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/IdempotencyKeyHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCurrentProfileRequest'
      responses:
        '200':
          $ref: '#/components/responses/CurrentUserSuccess'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token carried only in the Authorization header.
    JoinCapability:
      type: apiKey
      in: header
      name: X-Join-Capability
      description: Short-lived one-time capability bound to one ClassSection. The plaintext value is never logged or persisted; no ordinary Access Token is accepted by the Join operation.
  parameters:
    RequestIdHeader:
      name: X-Request-ID
      in: header
      required: false
      description: Optional caller correlation ID; the server returns the accepted/generated value in response header and meta.requestId.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: '^[A-Za-z0-9._:-]+$'
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Reuse for retries of the same normalized request; reuse with different content returns a conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[\x21-\x7E]+$'
...
  responses:
...
    SystemModeSuccess:
      description: Safe public SystemMode projection
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/SystemModeEnvelope' }
...
    CurrentUserSuccess:
      description: Current user and role profile projection
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/CurrentUserEnvelope' }
...
    Unauthorized:
      $ref: '#/components/responses/Error401'
    Forbidden:
      $ref: '#/components/responses/Error403'
...
    Conflict:
      $ref: '#/components/responses/Error409'
    UnprocessableEntity:
      $ref: '#/components/responses/Error422'
...
    InternalError:
      $ref: '#/components/responses/Error500'
    ServiceUnavailable:
      $ref: '#/components/responses/Error503'
...
    Error401:
      description: Unauthorized
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error403:
      description: Forbidden
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error409:
      description: Conflict
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error422:
      description: Unprocessable entity
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    Error500:
      description: Internal server error
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Error503:
      description: Service unavailable, maintenance, or read-only mode
      headers:
        X-Request-ID: { $ref: '#/components/headers/RequestIdResponseHeader' }
        Retry-After:
          description: Seconds until a retry may be attempted when known.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  headers:
    RequestIdResponseHeader:
      description: Server-accepted/generated request ID; equals meta.requestId on success and requestId on error.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 64
  schemas:
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 64
      pattern: '^[A-Za-z0-9._:-]+$'
      description: Opaque internal ID; clients must not parse prefixes or ordering.
...
    PositiveVersion:
      type: integer
      format: int64
      minimum: 1
...
    UserRole:
      type: string
      enum: [STUDENT, TEACHER, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, LOCKED, DISABLED]
...
    Gender:
      type: string
      enum: [MALE, FEMALE, OTHER]
...
    SystemMode:
      type: string
      enum: [NORMAL, READ_ONLY, MAINTENANCE]
...
    FieldError:
      type: object
      additionalProperties: false
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        i18nKey:
          type: string
          minLength: 1
          maxLength: 300
        params:
          type: object
          additionalProperties: true
      required: [field, code, i18nKey, params]
    ErrorCode:
      type: string
      description: Must be a standard error code from 07-enums-and-errors.md section 8; section 9 legacy aliases are not emitted.
      enum:
        - AUTH_REQUIRED
        - AUTH_CREDENTIAL_INVALID
        - AUTH_VERIFICATION_CODE_INVALID
        - AUTH_TOKEN_INVALID
        - AUTH_TOKEN_EXPIRED
        - AUTH_SESSION_REVOKED
        - AUTH_JOIN_CAPABILITY_INVALID
        - AUTH_JOIN_CAPABILITY_EXPIRED
        - AUTH_JOIN_CAPABILITY_ALREADY_USED
        - AUTH_ACCOUNT_DISABLED
        - AUTH_RATE_LIMITED
        - USER_NOT_FOUND
        - USER_IDENTITY_CONFLICT
        - USER_PROFILE_INVALID
        - USER_STATUS_NOT_ACTIVE
        - COURSE_NOT_FOUND
        - COURSE_CLASS_SECTION_NOT_FOUND
        - COURSE_CLASS_SECTION_NOT_JOINABLE
        - COURSE_CLASS_SECTION_NOT_WRITABLE
        - COURSE_SEMESTER_ARCHIVED
        - COURSE_DEADLINE_PASSED
        - COURSE_CHECKIN_WINDOW_CLOSED
        - COURSE_INVITE_INVALID
        - COURSE_INVITE_EXPIRED
        - COURSE_INVITE_REVOKED
        - COURSE_WRITE_DISABLED
        - COURSE_TEACHER_ASSIGNMENT_CONFLICT
        - ENROLLMENT_NOT_FOUND
        - ENROLLMENT_NOT_ACTIVE
        - ENROLLMENT_ALREADY_ACTIVE
        - ENROLLMENT_SEMESTER_CONFLICT
        - ENROLLMENT_TRANSITION_NOT_ALLOWED
        - ENROLLMENT_WITHDRAWAL_DISABLED
        - ENROLLMENT_REJOIN_DISABLED
        - ENROLLMENT_HAS_BLOCKING_WORK
        - ROSTER_IMPORT_NOT_FOUND
        - ROSTER_FILE_INVALID
        - ROSTER_SCHEMA_INVALID
        - ROSTER_IMPORT_DUPLICATE
        - ROSTER_IMPORT_NOT_READY
        - ROSTER_IMPORT_FAILED
        - ROSTER_IMPORT_SOURCE_UNSUPPORTED
        - ROSTER_ALIGNMENT_IN_PROGRESS
        - ROSTER_ALIGNMENT_SNAPSHOT_STALE
        - ROSTER_ALIGNMENT_INPUT_VERSION_CONFLICT
        - ROSTER_ALIGNMENT_EXCEPTION
        - ROSTER_RESOLUTION_INVALID
        - ROSTER_ALIGNMENT_RESULT_SUPERSEDED
        - ROSTER_IGNORE_NOT_ALLOWED
        - ROSTER_RESOLUTION_EVIDENCE_REQUIRED
        - SESSION_NOT_FOUND
        - SESSION_ALREADY_ACTIVE
        - SESSION_OUTSIDE_TIME_WINDOW
        - SESSION_TRANSITION_NOT_ALLOWED
        - SESSION_DURATION_CAP_REACHED
        - SESSION_ALREADY_COMPLETED
        - SESSION_ALREADY_USED
        - SESSION_NOT_COMPLETED
        - SESSION_EXPIRATION_NOT_ALLOWED
        - SESSION_RESUME_WINDOW_EXPIRED
        - SESSION_TIMELINE_INVALID
        - SESSION_EVENT_OUT_OF_ORDER
        - SESSION_RECONCILIATION_REQUIRED
        - EXERCISE_RECORD_NOT_FOUND
        - EXERCISE_RECORD_ALREADY_EXISTS_FOR_SESSION
        - EXERCISE_RECORD_DUPLICATE_SUBMISSION
        - EXERCISE_RECORD_DURATION_NOT_CREDITABLE
        - EXERCISE_RECORD_MEDIA_INCOMPLETE
        - EXERCISE_RECORD_DAILY_LIMIT_REACHED
        - EXERCISE_RECORD_WITHDRAWAL_NOT_ALLOWED
        - MEDIA_EVIDENCE_REQUIRED
        - MEDIA_COUNT_LIMIT_EXCEEDED
        - MEDIA_SIZE_EXCEEDED
        - MEDIA_TYPE_NOT_ALLOWED
        - MEDIA_CAPTURE_SOURCE_NOT_ALLOWED
        - MEDIA_UPLOAD_SESSION_EXPIRED
        - MEDIA_OBJECT_NOT_FOUND
        - MEDIA_INTEGRITY_MISMATCH
        - MEDIA_BIND_TARGET_INVALID
        - MEDIA_ALREADY_BOUND
        - MEDIA_PURPOSE_MISMATCH
        - MEDIA_NOT_AVAILABLE
        - MEDIA_ACCESS_DENIED
        - MEDIA_BOUND_TO_IMMUTABLE_RECORD
        - MEDIA_HAS_ACTIVE_BINDING
        - MEDIA_PROCESSING_ALREADY_STARTED
        - MEDIA_PROCESSING_INCOMPLETE
        - MEDIA_VERIFICATION_INCOMPLETE
        - MEDIA_TRANSITION_NOT_ALLOWED
        - MEDIA_FAILURE_NOT_RETRYABLE
        - MEDIA_RETENTION_HOLD
        - EXEMPTION_APPLICATION_NOT_FOUND
        - EXEMPTION_APPLICATION_TRANSITION_NOT_ALLOWED
        - EXEMPTION_APPLICATION_MEDIA_INVALID
        - PERMISSION_EXEMPTION_REVIEW_SCOPE_DENIED
        - REVIEW_NOT_FOUND
        - REVIEW_ALREADY_INITIALIZED
        - REVIEW_ALREADY_STARTED
        - REVIEW_ALREADY_COMPLETED
        - REVIEW_RESULT_REQUIRED
        - REVIEW_INVALID_REASON_REQUIRED
        - REVIEW_CHANGE_NOT_ALLOWED
        - REVIEW_BATCH_ITEM_FAILED
        - REVIEW_CREDIT_OVERRIDE_NOT_APPROVED
        - REVIEW_CREDIT_DURATION_INVALID
        - SCORE_NOT_FOUND
        - SCORE_ALREADY_EXISTS
        - SCORE_RULE_NOT_FOUND
        - SCORE_RULE_NOT_CONFIGURED
        - SCORE_RULE_APPROVAL_REQUIRED
        - SCORE_RULE_SELF_APPROVAL_NOT_ALLOWED
        - SCORE_RULE_DISTINCT_APPROVER_REQUIRED
        - SCORE_SOURCE_DATA_INCONSISTENT
        - SCORE_INPUT_INCOMPLETE
        - SCORE_INPUT_INVALID
        - SCORE_INPUT_VERSION_CONFLICT
        - SCORE_SOURCE_NOT_CHANGED
        - SCORE_ADJUSTMENT_INVALID
        - SCORE_ADJUSTMENT_NOT_ALLOWED
        - SCORE_ADJUSTMENT_APPROVAL_REQUIRED
        - SCORE_ADJUSTMENT_SELF_APPROVAL_NOT_ALLOWED
        - SCORE_ADJUSTMENT_EVIDENCE_INVALID
        - SCORE_NOT_PUBLISHABLE
        - SCORE_LOCKED
        - SCORE_CORRECTION_NOT_ALLOWED
        - VALIDATION_FAILED
        - VALIDATION_FIELD_REQUIRED
        - VALIDATION_ENUM_UNSUPPORTED
        - VALIDATION_DURATION_INVALID
        - VALIDATION_FORMAT_INVALID
        - PERMISSION_DENIED
        - PERMISSION_RESOURCE_NOT_FOUND
        - PERMISSION_RESOURCE_SCOPE_DENIED
        - PERMISSION_COURSE_SCOPE_DENIED
        - PERMISSION_REVIEW_SCOPE_DENIED
        - PERMISSION_AUDIT_SCOPE_DENIED
        - CONFLICT_STATE_TRANSITION
        - CONFLICT_VERSION_MISMATCH
        - CONFLICT_IDEMPOTENCY_KEY_REUSED
        - CONFLICT_REQUEST_IN_PROGRESS
        - CONFLICT_RESOURCE_ALREADY_EXISTS
        - CONFLICT_UNSUPPORTED_RESOURCE_STATE
        - SYSTEM_INTERNAL_ERROR
        - SYSTEM_SERVICE_UNAVAILABLE
        - SYSTEM_READ_ONLY
        - SYSTEM_MAINTENANCE
        - SYSTEM_MODE_UNSUPPORTED
        - SYSTEM_DATA_INTEGRITY_ERROR
        - SYSTEM_DEPENDENCY_TIMEOUT
        - AUDIT_WRITE_FAILED
        - AUDIT_RETENTION_POLICY_REQUIRED
    ErrorDetails:
      type: object
      additionalProperties: false
      properties:
        fieldErrors:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/FieldError' }
        resourceType:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        resourceId:
          type: string
          minLength: 1
          maxLength: 64
        currentState:
          type: string
          pattern: '^[A-Z][A-Z0-9_]*$'
        allowedActions:
          type: array
          maxItems: 100
          uniqueItems: true
          items: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        expectedVersion:
          type: integer
          minimum: 0
        actualVersion:
          type: integer
          minimum: 0
        retryAfterSeconds:
          type: integer
          minimum: 0
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 128
        itemErrors:
          type: array
          maxItems: 100
          items: { $ref: '#/components/schemas/BatchItemError' }
        migrationReference:
          type: string
          minLength: 1
          maxLength: 256
    BatchItemError:
      type: object
      additionalProperties: false
      properties:
        itemId:
          type: string
          minLength: 1
          maxLength: 128
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
        details: { $ref: '#/components/schemas/ErrorDetails' }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        timestamp:
          type: string
          format: date-time
      required: [itemId, code, message, details, requestId, timestamp]
...
    SuccessMeta:
      type: object
      additionalProperties: false
      properties:
        requestId:
          type: string
          minLength: 1
          maxLength: 64
      required: [requestId]
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message:
          type: string
          minLength: 1
          maxLength: 1000
        details: { $ref: '#/components/schemas/ErrorDetails' }
        requestId:
          type: string
          minLength: 1
          maxLength: 64
        timestamp:
          type: string
          format: date-time
      required: [code, message, details, requestId, timestamp]
      example:
        code: CONFLICT_VERSION_MISMATCH
        message: The resource changed. Refresh and try again.
        details:
          resourceType: EXERCISE_RECORD
          resourceId: rec_01JABC123
          expectedVersion: 3
          actualVersion: 4
        requestId: req_01JABC123
        timestamp: '2026-08-02T12:00:00Z'
...
    SystemModeProjection:
      type: object
      additionalProperties: false
      properties:
        mode: { $ref: '#/components/schemas/SystemMode' }
        policyVersion: { $ref: '#/components/schemas/PositiveVersion' }
        updatedAt: { type: string, format: date-time }
      required: [mode, policyVersion, updatedAt]
    SystemModeEnvelope:
      type: object
      additionalProperties: false
      properties:
        data: { $ref: '#/components/schemas/SystemModeProjection' }
        meta: { $ref: '#/components/schemas/SuccessMeta' }
      required: [data, meta]
...
    User:
      type: object
      additionalProperties: false
      description: Minimal authorized projection; authentication internals and plaintext contact methods are never returned.
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        primaryEmailMasked: { type: [string, 'null'], maxLength: 254, readOnly: true, description: Masked projection only; plaintext email is never returned. }
        primaryPhoneMasked: { type: [string, 'null'], maxLength: 32, readOnly: true, description: Masked projection only; plaintext phone is never returned. }
        emailVerified: { type: boolean, readOnly: true, description: Derived from the current contact verification fact; the verification timestamp is not exposed. }
        phoneVerified: { type: boolean, readOnly: true, description: Derived from the current contact verification fact; the verification timestamp is not exposed. }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, role, status, primaryEmailMasked, primaryPhoneMasked, emailVerified, phoneVerified, version]
    StudentProfile:
      type: object
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        userId: { $ref: '#/components/schemas/OpaqueId' }
        studentNumber: { type: string, minLength: 1, maxLength: 32 }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        gender: { $ref: '#/components/schemas/Gender' }
        gradeYear: { type: integer, minimum: 2000, maximum: 2027, description: Contract-year bound; update annually and enforce the current-year-plus-one rule at runtime. }
        collegeName: { type: [string, 'null'], maxLength: 200 }
        majorName: { type: [string, 'null'], maxLength: 200 }
        administrativeClassName: { type: [string, 'null'], maxLength: 200 }
        status: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, userId, studentNumber, fullName, gender, gradeYear, collegeName, majorName, administrativeClassName, status, createdAt, updatedAt, deletedAt, version]
    TeacherProfile:
      type: object
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        userId: { $ref: '#/components/schemas/OpaqueId' }
        employeeNumber: { type: string, minLength: 1, maxLength: 32 }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        collegeName: { type: [string, 'null'], maxLength: 200 }
        departmentName: { type: [string, 'null'], maxLength: 200 }
        title: { type: [string, 'null'], maxLength: 100 }
        status: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, userId, employeeNumber, fullName, status, createdAt, updatedAt, deletedAt, version]
    AdminProfile:
      type: object
      additionalProperties: false
      properties:
        id: { $ref: '#/components/schemas/OpaqueId' }
        organizationId: { $ref: '#/components/schemas/OpaqueId' }
        userId: { $ref: '#/components/schemas/OpaqueId' }
        employeeNumber: { type: string, minLength: 1, maxLength: 32 }
        fullName: { type: string, minLength: 1, maxLength: 100 }
        departmentName: { type: [string, 'null'], maxLength: 200 }
        status: { type: string, pattern: '^[A-Z][A-Z0-9_]*$' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        deletedAt: { type: [string, 'null'], format: date-time }
        version: { $ref: '#/components/schemas/PositiveVersion' }
      required: [id, organizationId, userId, employeeNumber, fullName, status, createdAt, updatedAt, deletedAt, version]
...
    CurrentUserData:
      type: object
      additionalProperties: false
      required: [user, studentProfile, teacherProfile, adminProfile]
      properties:
        user: { $ref: '#/components/schemas/User' }
        studentProfile: { oneOf: [{ $ref: '#/components/schemas/StudentProfile' }, { type: 'null' }] }
        teacherProfile: { oneOf: [{ $ref: '#/components/schemas/TeacherProfile' }, { type: 'null' }] }
        adminProfile: { oneOf: [{ $ref: '#/components/schemas/AdminProfile' }, { type: 'null' }] }
...
    UpdateCurrentProfileRequest:
      type: object
      additionalProperties: false
      anyOf:
        - required: [primaryEmail]
        - required: [primaryPhone]
      properties:
        primaryEmail: { type: [string, 'null'], format: email, maxLength: 254 }
        primaryPhone: { type: [string, 'null'], maxLength: 32 }
        expectedVersion: { $ref: '#/components/schemas/PositiveVersion' }
      required: [expectedVersion]
      description: Contact-change transport placeholder; it must fail closed with SYSTEM_MODE_UNSUPPORTED until a verified contact-change flow is approved.
...
    CurrentUserEnvelope: { $ref: '#/components/schemas/CurrentUserResponse' }
...
    CurrentUserResponse:
      type: object
      required: [data, meta]
      properties: { data: { $ref: '#/components/schemas/CurrentUserData' }, meta: { $ref: '#/components/schemas/SuccessMeta' } }
...